Move FFE exposure logic into datadog-ffe
