+++ datadog-ffe/Cargo.toml
@@ -24,7 +24,9 @@ serde-bool = { version = "0.1.3", default-features = false }
 serde_with = { version = "3.11.0", default-features = false, features = ["base64", "hex", "macros"] }
 thiserror = { version = "2.0.3", default-features = false }
 url = { version = "2.5.0", default-features = false, features = ["std"] }
+lru = { version = "0.16.3", optional = true }
 pyo3 = { version = "0.28", optional = true, default-features = false, features = ["macros"] }
 
 [features]
+telemetry = ["dep:lru"]
 pyo3 = ["dep:pyo3"]
